pyproject.toml
[tool.poetry]
name = "ubnetdeforchestrator"
version = "0.1.0"
description = ""
authors = ["Blake <80996688+BlakeTnr@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
proxmoxer = "^2.2.0"
pytest = "^8.4.1"
requests = "^2.32.4"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
